requirements.txt
datetime-utils @ git+https://github.com/reimibeta/django-datetime-utils.git@545cd2e124497406647b538a3a371549aaca4b8c
Django==3.2.3
django-admin-list-filter-dropdown==1.0.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.0
drf-flex-fields==0.9.0
rest-framework-utils @ git+https://github.com/reimibeta/django-rest-framework-utils.git@764007b810f1c3609cca6c0e146da746531aa929